// types/api.tsx
export type ApiResponse<T = any> = {
  data?: T[];
  links?: Links;
  meta?: Meta;
  success: boolean;
  status: number;
};
// ...
type RatingProgress = {
  stars: number;
  count: number;
  percentage: number;
};

type RatingStats = {
  total_reviews: number;
  average_rating: number;
  rating_progress: RatingProgress[];
};
// ...
export type ProductDetails = {
  id: number;
  name: string;
  slug: string;
  stocks: [];
  slugData: string;
  added_by: string;
  seller_id: number;
  shop_id: number;
  shop_name: string;
  shop_logo: string;
  photos?: Array<{
    variant?: string;
    path?: string;
  }>;
  thumbnail_image: string;
  tags: string[];
  price_high_low: string;
  choice_options: any[];
  colors: any[];
  has_discount: boolean;
  discount: number;
  stroked_price: string;
  main_price: string;
  calculable_price: number;
  currency_symbol: string;
  current_stock: number;
  unit: string;
  rating: number;
  rating_count: number;
  rating_stats: RatingStats;
  earn_point: number;
  description: string;
  short_description: string;
  ingredient: string;
  video_link: string;
  brand: {
    id: number;
    name: string;
    logo: string;
  };
  link: string;
  meta_title: string;
  meta_description: string;
  meta_img: string;
  discount_text: string;
}
// ...
export interface Links {
  first: string;
  last: string;
  prev: string | null;
  next: string | null;
}

// Define the type for the pagination link item
export interface MetaLink {
  url: string | null;
  label: string;
  active: boolean;
}

// Define the type for the meta information
export interface Meta {
  current_page: number;
  from: number;
  last_page: number;
  links: MetaLink[]; // Array of MetaLink items
  path: string;
  per_page: number;
  to: number;
  total: number;
}